Move async code to async/libasync.la
