+++ src/EventLoop.cc
@@ -9,7 +9,7 @@
 /* DEBUG: section 01    Main Loop */
 
 #include "squid.h"
-#include "AsyncEngine.h"
+#include "async/AsyncEngine.h"
 #include "base/AsyncCallQueue.h"
 #include "debug/Stream.h"
 #include "EventLoop.h"
