+++ src/adaptation/ecap/ServiceRep.cc
@@ -13,7 +13,7 @@
 #include "adaptation/ecap/Host.h"
 #include "adaptation/ecap/ServiceRep.h"
 #include "adaptation/ecap/XactionRep.h"
-#include "AsyncEngine.h"
+#include "async/AsyncEngine.h"
 #include "base/TextException.h"
 #include "debug/Stream.h"
 #include "EventLoop.h"
