+++ src/async/AsyncEngine.cc
@@ -7,5 +7,4 @@
  */
 
 #include "squid.h"
-#include "AsyncEngine.h"
-
+#include "async/AsyncEngine.h"
