+++ src/async/Makefile.am
@@ -1,0 +1,14 @@
+## Copyright (C) 1996-2023 The Squid Software Foundation and contributors
+##
+## Squid software is distributed under GPLv2+ license and includes
+## contributions from numerous individuals and organizations.
+## Please see the COPYING and CONTRIBUTORS files for details.
+##
+
+include $(top_srcdir)/src/Common.am
+
+noinst_LTLIBRARIES = libasync.la
+
+libasync_la_SOURCES = \
+	AsyncEngine.cc \
+	AsyncEngine.h
