+++ src/tests/testEventLoop.cc
@@ -7,7 +7,7 @@
  */
 
 #include "squid.h"
-#include "AsyncEngine.h"
+#include "async/AsyncEngine.h"
 #include "compat/cppunit.h"
 #include "EventLoop.h"
 #include "time/Engine.h"
